src/functions/store_html/requirements.txt
functions-framework==3.5.0
google-cloud-storage==2.18.2
google-cloud-pubsub==2.23.1
google-cloud-firestore==2.16.0
apify-client==1.8.0
google-cloud-logging==3.10.0
google-cloud-secret-manager==2.20.2